refactor: improve chat exception naming and correct their location
