+++ chat/src/main/kotlin/net/thechance/chat/api/exception/ChatException.kt
@@ -1,0 +1,9 @@
+package net.thechance.chat.api.exception
+
+
+open class ChatException(
+    val code: Int?= null,
+    override val message: String,
+) : Exception(message)
+
+class NotFoundException(message: String) : ChatException(message = message)
