--- chat/src/main/kotlin/net/thechance/chat/exception/ChatException.kt
@@ -1,9 +1,0 @@
-package net.thechance.chat.exception
-
-import org.springframework.http.HttpStatus
-
-open class ChatException(
-    val code: Int,
-    val status: HttpStatus,
-    override val message: String,
-) : Exception(message)
