--- chat/src/main/kotlin/net/thechance/chat/service/exception/ChatException.kt
@@ -1,3 +1,0 @@
-package net.thechance.chat.service.exception
-
-class NotFoundException(message: String) : RuntimeException(message)
